feat: implement PaymentService for handling recurring transactions and payment notifications
